pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "nested_ragged_tensors"
version = "0.0.1"
authors = [
  { name="Matthew McDermott", email="mattmcdermott8@gmail.com" },
]
description = "Utilities for efficiently working with, saving, and loading, collections of connected nested ragged tensors in PyTorch"
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = ["safetensors", "numpy"]

[project.optional-dependencies]
dev = ["pre-commit"]
tests = ["pytest", "ml-mixins", "loguru", "torch", "rootutils", "hydra-core", "humanize"]

[project.urls]
Homepage = "https://github.com/mmcdermott/nested_ragged_tensors"
Issues = "https://github.com/mmcdermott/nested_ragged_tensors/issues"
